FloatArray debug variable type for calibration (#1164)
* "trace implementation cleanup; updates #1154"

* "variable length encoded, but only used trivially, errors in debug protocol; updates #1154"

* "trace handler and circular buffer errors fixed; updates #1154"

* "disabled obsolete trace handler test; updates #1154"

* "FloatArray debug var type and tests; updates #1154"

* "float array implemented and debug interface reads values; updates #1154"

* "FloatArray can be manipulated via debug interface; updates #1154"

* "each pinch valve keeps its own calibration, initial values hard-coded; updates #1154"

* "fixed integration test bug, improved FloatArray test; updates #1154"

* "removing redundant gitignore; updates #1154"

* "rule to ignore rendudant gitignore; updates #1154"

* "refactoring, documentation and more tests for Trace and its command handler, fixes in python interface; updates #1154"

* "more cleanup and warnings; updates #1154"

* "reverting overzealous replace; updates #1154"

* "addressed PR comments; updates #1154"

+++ software/controller/lib/hal/circular_buffer.h
@@ -15,8 +15,7 @@ limitations under the License.
 
 #pragma once
 
-#include <stdint.h>
-
+#include <cstdint>
 #include <optional>
 
 #include "hal.h"
@@ -27,7 +26,7 @@ limitations under the License.
 // and the interrupt handlers, so it needs to be thread safe.
 // I'm disabling interrupts during the critical sections to
 // ensure that's the case.
-template <class T, uint N>
+template <class T, size_t N>
 class CircularBuffer {
   // Uses an array of size N+1 to hold all N elements of the buffer, as
   // buffer_[head_] is by definition inaccessible.
@@ -37,20 +36,21 @@ class CircularBuffer {
 
  public:
   CircularBuffer() {
-    for (uint i = 0; i <= N; ++i) buffer_[i] = T();
+    for (size_t i = 0; i <= N; ++i) buffer_[i] = T();
   }
 
   // Return number of elements available in the buffer to read.
-  int FullCount() const {
+  size_t FullCount() const {
     BlockInterrupts block;
-    int ct = head_ - tail_;
+    /// \TODO: ssize_t is likely compiler-dependent; this function could be improved
+    ssize_t ct = head_ - tail_;
     if (ct < 0) ct += N + 1;
-    return ct;
+    return static_cast<size_t>(ct);
   }
 
   // Return number of free spaces in the buffer where more
   // elements can be written.
-  int FreeCount() const { return N - FullCount(); }
+  size_t FreeCount() const { return N - FullCount(); }
 
   // Get the oldest element from the buffer, popping it from the buffer.
   std::optional<T> Get() {
